+++ Source/Objects/Hardware/IPE/IpeV4 FLT/ORIpeV4FLTModel.m
@@ -211,7 +211,7 @@ static IpeRegisterNamesStruct regV4[kFLTV4NumRegs] = {
 	return (eventMask & (1L<<chan)) != 0;
 }
 
-- (int) stationNumber
+- (NSUInteger) stationNumber
 {
 	//is it a minicrate?
 	if([[[self crate] class]  isSubclassOfClass: NSClassFromString(@"ORIpeV4MiniCrateModel")]){
